Rebase on solver move PR.
