+++ backends/p4tools/common/lib/ir_compare.h
@@ -25,25 +25,5 @@ struct StateVariableLess {
     }
 };
 
-/// Equals for SymbolicVariable pointers. We only compare the label.
-struct SymbolicVariableEqual {
-    bool operator()(const IR::SymbolicVariable *s1, const IR::SymbolicVariable *s2) const {
-        return s1->label == s2->label;
-    }
-    bool operator()(const IR::SymbolicVariable &s1, const IR::SymbolicVariable &s2) const {
-        return s1.label == s2.label;
-    }
-};
-
-/// Less for SymbolicVariable pointers. We only compare the label.
-struct SymbolicVariableLess {
-    bool operator()(const IR::SymbolicVariable *s1, const IR::SymbolicVariable *s2) const {
-        return *s1 < *s2;
-    }
-    bool operator()(const IR::SymbolicVariable &s1, const IR::SymbolicVariable &s2) const {
-        return s1 < s2;
-    }
-};
-
 }  // namespace IR
 #endif /* BACKENDS_P4TOOLS_COMMON_LIB_IR_COMPARE_H_ */
