+++ ir/compare.h
@@ -1,0 +1,29 @@
+#ifndef IR_COMPARE_H_
+#define IR_COMPARE_H_
+
+#include "ir/ir.h"
+
+namespace IR {
+
+/// Equals for SymbolicVariable pointers. We only compare the label.
+struct SymbolicVariableEqual {
+    bool operator()(const IR::SymbolicVariable *s1, const IR::SymbolicVariable *s2) const {
+        return s1->label == s2->label;
+    }
+    bool operator()(const IR::SymbolicVariable &s1, const IR::SymbolicVariable &s2) const {
+        return s1.label == s2.label;
+    }
+};
+
+/// Less for SymbolicVariable pointers. We only compare the label.
+struct SymbolicVariableLess {
+    bool operator()(const IR::SymbolicVariable *s1, const IR::SymbolicVariable *s2) const {
+        return *s1 < *s2;
+    }
+    bool operator()(const IR::SymbolicVariable &s1, const IR::SymbolicVariable &s2) const {
+        return s1 < s2;
+    }
+};
+
+}  // namespace IR
+#endif /* IR_COMPARE_H_ */
